ExperimentManager file added, minor tweaks to ExperimentClient
