+++ mlflow/src/tracking_server/ExperimentClient.js
@@ -8,8 +8,7 @@ class ExperimentClient {
    * Validates that another experiment with the same name does not already exist and fails if another experiment with the same name already exists.
    *
    * @param {string} name Experiment name.  (required)
-   * @param {string} artifact_location Optional location where all artifacts for the experiment are stored.
-   *    If not provided, the remote server will select an appropriate default.
+   * @param {string} artifact_location Optional location where all artifacts for the experiment are stored.  If not provided, the remote server will select an appropriate default.
    * @param {Array<{key: string, value: string}>} tags Optional collection of tags to set on the experiment.
    * @returns {Promise<Object>} Returns the ID of the newly created experiment in an object.
    */
@@ -40,13 +39,11 @@ class ExperimentClient {
   /**
    * Search experiments.
    *
-   * @param {string} filter A filter expression over experiment attributes and tags that allows returning a subset of experiments
-   *    The syntax is a subset of SQL.  (required)
+   * @param {string} filter A filter expression over experiment attributes and tags that allows returning a subset of experiments.  The syntax is a subset of SQL.  (required)
    * @param {int64} max_results Maximum number of experiments desired.  (required)
    * @param {string} page_token Optional token indicating the page of experiments to fetch.
    * @param {Array<string>} order_by Optional list of columns for ordering search results.
-   * @param {string} view_type Optional qualifier for type of experiments to be returned.
-   *    See https://mlflow.org/docs/latest/rest-api.html#mlflowviewtype
+   * @param {string} view_type Optional qualifier for type of experiments to be returned.  See https://mlflow.org/docs/latest/rest-api.html#mlflowviewtype
    * @returns {Promise<Object>} Returns object containing an array of experiment objects matching the filter,
    *    and optionally a next_page_token that can be used to retrieve the next page of experiments.
    */
@@ -280,230 +277,3 @@ class ExperimentClient {
 }
 
 export { ExperimentClient };
-
-
-
-
-
-
-
-
-
-  /**
-   * Full workflow of creating, naming and starting a run under an experiment (referenced by ID),
-   * logging metrics, params, and tags, logging the model, and finishing the run.
-   *
-   * @param {string} experiment_id ID of the experiment under which to log the run.  (required)
-   * @param {string} run_name Name of the run to be created and run (optional)
-   * @param {Array<{key: string, value: number, timestamp: number, step?: number}>} [metrics] The metrics to log (up to 1000 metrics) (optional)
-   * @param {Array<{key: string, value: string}>} [params] The params to log (up to 100 params) (optional)
-   * @param {Array<{key: string, value: string}>} [tags=[]] The tags to log (up to 100 tags) (optional)
-   * @param {Object} model The ML model data to log to the run, represented as a Javascript object
-   * @returns {Promise<Object>} The created run object with updated metadata
-   */
-  async function withStartExperimentRunByExperimentId(
-    experiment_id,
-    run_name = null,
-    metrics = [],
-    params = [],
-    tags = [],
-    model
-  ) {
-    if (!experiment_id) {
-      throw new Error('Experiment ID is required');
-    }
-  
-    // create run
-    const run = await runManagement.createRun(experiment_id, run_name, tags);
-    const run_id = run.info.run_id;
-  
-    // log metric, params, and tags via logBatch
-    await runManagement.logBatch(run_id, metrics, params, tags);
-  
-    // log model
-    // (model gets passed in as a JS object, not JSON - it gets JSON stringified here after adding a run_id property)
-    model.run_id = run_id;
-    let model_json = JSON.stringify(model);
-    await runManagement.logModel(run_id, model_json);
-  
-    // updateRun to finish it
-    const latestRun = await runManagement.updateRun(run_id, 'FINISHED');
-  
-    return latestRun;
-  }
-  
-  // test ****************************************************************************************************************************************
-  const testWithStartExperimentRunByExperimentId = async () => {
-    const metrics = [
-      { key: 'metric1', value: 0.111, timestamp: Date.now() },
-      { key: 'metric2', value: 0.222, timestamp: Date.now() },
-    ];
-    const params = [
-      { key: 'testParam', value: 'testParamValue' },
-      { key: 'testParam2', value: 'testParamValue2' },
-    ];
-    const tags = [
-      { key: 'testKey', value: 'testValue' },
-      { key: 'testKey2', value: 'testValue2' },
-    ];
-    const model = {
-      artifact_path: 'model',
-      flavors: {
-        python_function: {
-          model_path: 'model.pkl',
-          loader_module: 'mlflow.sklearn',
-          python_version: '3.8.10',
-        },
-      },
-      model_url: 'STRING',
-      model_uuid: 'STRING',
-      utc_time_created: Date.now(),
-      mlflow_version: 'STRING',
-    };
-    const log = await withStartExperimentRunByExperimentId(
-      '977566317259111173',
-      undefined,
-      metrics,
-      params,
-      tags,
-      model
-    );
-    return console.log(log);
-  };
-  // uncomment below ---
-  // testWithStartExperimentRunByExperimentId();
-  
-  /**
-   * Full workflow of creating, naming and starting a run under an experiment (referenced by name),
-   * logging metrics, params, and tags, logging the model, and finishing the run.
-   *
-   * @param {string} experiment_name Name of the experiment under which to log the run.  (required)
-   * @param {string} run_name Name of the run to be created and run (optional)
-   * @param {Array<{key: string, value: number, timestamp: number, step?: number}>} [metrics] The metrics to log (up to 1000 metrics) (optional)
-   * @param {Array<{key: string, value: string}>} [params] The params to log (up to 100 params) (optional)
-   * @param {Array<{key: string, value: string}>} [tags=[]] The tags to log (up to 100 tags) (optional)
-   * @param {Object} model The ML model data to log to the run, represented as a Javascript object
-   * @returns {Promise<Object>} The created run object with updated metadata
-   */
-  async function withStartExperimentRunByExperimentName(
-    experiment_name,
-    run_name = null,
-    metrics = [],
-    params = [],
-    tags = [],
-    model
-  ) {
-    if (!experiment_name) {
-      throw new Error('Experiment name is required');
-    }
-  
-    let experiment_id;
-    try {
-      const exp = await getExperimentByName(experiment_name);
-      experiment_id = exp.experiment_id;
-    } catch {
-      experiment_id = await createExperiment(experiment_name);
-    }
-  
-    // create run
-    const run = await runManagement.createRun(experiment_id, run_name, tags);
-    const run_id = run.info.run_id;
-  
-    // log metric, params, and tags via logBatch
-    await runManagement.logBatch(run_id, metrics, params, tags);
-  
-    // log model
-    // (model gets passed in as a JS object, not JSON - it gets JSON stringified here after adding a run_id property)
-    model.run_id = run_id;
-    let model_json = JSON.stringify(model);
-    await runManagement.logModel(run_id, model_json);
-  
-    // updateRun to finish it
-    const latest_run = await runManagement.updateRun(run_id, 'FINISHED');
-  
-    return latest_run;
-  }
-  
-  // test ****************************************************************************************************************************************
-  const testWithStartExperimentRunByExperimentName = async () => {
-    const metrics = [
-      { key: 'metric1', value: 0.9, timestamp: Date.now() },
-      { key: 'metric2', value: 0.222, timestamp: Date.now() },
-    ];
-    const params = [
-      { key: 'testParam', value: 'testParamValue' },
-      { key: 'testParam2', value: 'testParamValue2' },
-    ];
-    const tags = [
-      { key: 'testKey', value: 'testValue' },
-      { key: 'testKey2', value: 'testValue2' },
-    ];
-    const model = {
-      artifact_path: 'model',
-      flavors: {
-        python_function: {
-          model_path: 'model.pkl',
-          loader_module: 'mlflow.sklearn',
-          python_version: '3.8.10',
-        },
-      },
-      model_url: 'STRING',
-      model_uuid: 'STRING',
-      utc_time_created: Date.now(),
-      mlflow_version: 'STRING',
-    };
-    const log = await withStartExperimentRunByExperimentName(
-      'experimentSummary test',
-      undefined,
-      metrics,
-      params,
-      tags,
-      model
-    );
-    return console.log(log);
-  };
-  // uncomment below ---
-  // testWithStartExperimentRunByExperimentName();
-  
-  /**
-   * Returns an array of all the passed-in experiment's runs, sorted according to the passed-in metric
-   *
-   * @param {string} experiment_id The experiment whose runs will be evaluated (required)
-   * @param {string} primaryMetric The metric by which the results array will be sorted
-   * @param {string | number} order Sort order for the array: pass in 'DESC' or 1 for descending; 'ASC' or -1 for ascending
-   * @returns Promise<Array<Object>> An array of run objects belonging to the passed-in experiment ID, sorted according to the primary metric
-   */
-  async function experimentSummary(experiment_id, primaryMetric, order) {
-    // use Search Runs to return all runs whose experiment ID matches the passed in one
-    // use Search Runs's order_by field to sort results array by primaryMetric
-    let orderString;
-    if (order === 1 || order === 'DESC') orderString = 'DESC';
-    else if (order === -1 || order === 'ASC') orderString = 'ASC';
-    const arg = `metrics.${primaryMetric} ${orderString}`;
-    const data = await runManagement.searchRuns(
-      [experiment_id],
-      undefined,
-      undefined,
-      undefined,
-      [arg]
-    );
-    return data.runs;
-  }
-  
-  // test ****************************************************************************************************************************************
-  const testExperimentSummary = async () => {
-    const log = await experimentSummary('787867007534323476', 'metric1', 'ASC');
-    return console.log(log);
-  };
-  // uncomment below ---
-  // testExperimentSummary();
-
-
-
-
-
-const MLFLOW_TRACKING_URI = 'http://localhost:5001/api/2.0/mlflow';
-import { RunManagement } from './run_management.js';
-const path = 'artifacts';
-const runManagement = new RunManagement('http://localhost:5001', path);
-
